feat: refactor summarization logic
