+++ .gitignore
@@ -1,3 +1,6 @@
+# Release notes sample output folder
+review_release_notes/
+
 .DS_Store
 
 # Byte-compiled / optimized / DLL files
