+++ release_notes/input_repos.json
@@ -1,3 +1,3 @@
 [
-    "Ho1yShif/test-release-pipeline",
+    "Ho1yShif/test-release-pipeline"
 ]
